backend/requirements.txt
flask
flask-cors
google-generativeai
minio
nltk
python-dotenv